pyproject.toml
[project]
name = "background-task-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "django>=5.1.4",
    "djangorestframework>=3.15.2",
    "django-celery-results>=2.5.1",
    "python-decouple>=3.8",
    "redis>=5.2.1",
]
